requirements.txt
PyJWT>=1.4.0,<2.0.0
django
graphene-django
